restructured playlist code
